[TF][MAPS] support geo: intents
This patch adds the ability for Telegram to parse locations from send
intents containing a geo:<lat>,<lon>,<zoom> string.

This is done by various map applications, including the popular OSMAnd
and Orux maps. This patch enables the user to share the position of a
POI or destination with Telegram and have it send as a location.

(cherry picked from commit 30cdb45039672b2aaba29c6ffcd89752b3ca25c8)