Fix Notification Request popup
The original logic considered a denial before 3 seconds to be "accidental". Many users can read the popup and decide long before then, just to be asked again a day later.

I think it should be half a second (500), but 1 second is much better than 3

This should fix bug #752 , although the user was incorrect in their theory, they were just hitting it too quickly